% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterworth_filter.R
\name{butterworth_filter}
\alias{butterworth_filter}
\title{Butterworth filter}
\usage{
butterworth_filter(
  x,
  filter_type = "low",
  filter_order = 5,
  filter_cutoffs = 20,
  sampling_rate = 200
)
}
\arguments{
\item{x}{input signal to filter}

\item{filter_type}{Filter type; either "low", "high", "pass" or "stop" (default is "low")}

\item{filter_order}{Order of the filter (default is 5)}

\item{filter_cutoffs}{Cut-off frequency for low-pass filter (default is 20)}

\item{sampling_rate}{Sampling rate of signal (default is 200)}
}
\value{
Returns the filtered signal
}
\description{
Construct Butterworth filter and fit to signal with initial conditions.
}
\note{
: This function depends on 'gsignal' package as the 'signal' package implementations do a poor job with the initial state and lead to highly inaccurate signal edges. For example, gsignal::filter and gsignal::filtfilt give same results (only difference is that filtfilt gives 1D array as return, so doesn't need to index 'y' variable as is done inside this function)
}
