% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_difference.R
\name{central_difference}
\alias{central_difference}
\title{Calculate central difference of a signal}
\usage{
central_difference(x, sampling_rate, endpoints = TRUE, abs = FALSE)
}
\arguments{
\item{x}{numeric input}

\item{sampling_rate}{sampling rate in Hz (numeric)}

\item{endpoints}{calculate first and last values using backward/forward difference (default is TRUE)}

\item{abs}{return absolute values (default is FALSE)}
}
\description{
Calculate central difference of a signal
}
\note{
If \code{endpoints=TRUE} (default) then \code{x[1]} is calculated using backward_difference and \code{x[length(x)]} (last value in \code{x}) is calculated using the forward_difference. All other values are calculated using the central difference. If \code{endpoints=FALSE} then \code{x[1]} and \code{x[length(x)]} will be NA.
}
\examples{
\dontrun{
# create time object with 0.005s step size (sampling_rate=200)
t <- seq(0, 1, by=0.005)
# compute sine wave over t
y <- sin(2*pi*t)
# calculate central difference
central_difference(y, sampling_rate=200)
# calculate central difference with absolute values and plot against t
y_diff <- central_difference(y, sampling_rate=200, abs=TRUE)
plot(t, y_diff, type="l")
# do not calculate first/last values and instead return NA
central_difference(y, sampling_rate=200, endpoints=FALSE)
}
}
