% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr1_genes}
\alias{chr1_genes}
\title{Genomic feature example data}
\format{
An object of class \code{GRanges} of length 1456.
}
\source{
\url{https://www.ensembl.org}
}
\usage{
data(chr1_genes)
}
\value{
GRanges object with 1456 ranges and 2 metadata columns.
\describe{
\item{chr1_genes}{Ranges represent gene features for chromosome 1 of
zebrafish GRCz11 genome.}
}
}
\description{
Gene features for example usage. Generation of this data is
documented in \code{system.file("data-raw/chr1_genes.R", package = "tadar")}.
}
\keyword{datasets}
