% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr1_tt}
\alias{chr1_tt}
\title{Differential expression example data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 716 rows and 5 columns.
}
\usage{
data(chr1_tt)
}
\value{
A 716 x 5 tibble object.
}
\description{
\emph{p}-values from differential expressiong testing for
example usage.
}
\keyword{datasets}
