% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/convertRanges.R
\name{convertRanges}
\alias{convertRanges}
\alias{convertRanges,GRangesList-method}
\title{Convert ranges from single base positions to elastic windows,
or vice versa}
\usage{
convertRanges(dar, extendEdges = FALSE)

\S4method{convertRanges}{GRangesList}(dar, extendEdges = FALSE)
}
\arguments{
\item{dar}{A GRangesList with ranges representing single base positions
and metadata columns containing the smoothed DAR values.}

\item{extendEdges}{A logical specifying if ranges at the edges of each
chromosome should be extended to cover the entire chromosome.
This argument is only used when converting to ranges associated with the
smooth DAR values (i.e. \verb{to = "smooth}).
Useful for downstream assignment of DAR values to genomic features that
exist at the 5' or 3' edges of the chromosome.}
}
\value{
A GRangesList with ranges that represent the elastic sliding windows
used to smooth the DAR metric.
Each element of the list represents a comparison between two sample groups.
}
\description{
Convert the ranges element associated with single base position
raw DAR values to the elastic sliding window ranges associated with the
smoothed DAR values, or vice versa
}
\examples{
fl <- system.file("extdata", "chr1.vcf.bgz", package="darr")
genotypes <- readGenotypes(fl)
groups <- list(
    group1 = c("S2", "S7", "S9", "S10", "S19", "S20"),
    group2 = c("S3", "S6", "S11", "S12", "S15", "S16", "S18")
)
counts <- countAlleles(genotypes, groups)
props <- countsToProps(counts)
contrasts <- matrix(
    data = c(1, -1),
    dimnames = list(
        Levels = c("group1", "group2"),
        Contrasts = c("group1v2")
    )
)
dar <- dar(props, contrasts)
convertRanges(dar)

## Convert back to origin ranges
darWindows <- convertRanges(dar)
convertRanges(darWindows)

}
