% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_difference.R
\name{forward_difference}
\alias{forward_difference}
\title{Calculate forward difference of a signal}
\usage{
forward_difference(x, sampling_rate, replace_first = NA, abs = FALSE)
}
\arguments{
\item{x}{numeric input}

\item{sampling_rate}{sampling rate in Hz (numeric)}

\item{replace_first}{numeric value used to replace NA in the first cell (default is NA)}

\item{abs}{return absolute values (default is FALSE)}
}
\description{
Calculate forward difference of a signal
}
\note{
As foward_difference calculates the difference between points, it does not by default return a difference value for the first value of a signal, because it cannot be differenced against a value that occurs before it. Use central_difference for reliable differencing across the entire signal.
}
\examples{
\dontrun{
# create time object with 0.005s step size (sampling_rate=200)
t <- seq(0, 1, by=0.005)
# compute sine wave over t
y <- sin(2*pi*t)
# calculate forward difference
forward_difference(y, sampling_rate=200)
forward_difference(y, sampling_rate=200, abs=TRUE)
# replace first cell with the first value of y
forward_difference(y, sampling_rate=200, replace_first = dplyr::first(y))
}
}
