% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_velocity.R
\name{get_velocity}
\alias{get_velocity}
\title{Get velocity}
\usage{
get_velocity(x, sampling_rate)
}
\arguments{
\item{x}{signal to differentiate}

\item{sampling_rate}{sampling rate of signal}
}
\value{
Differenced signal
}
\description{
Calculate absolute-valued first difference of a signal.
}
\note{
This function is deprecated, use \code{tadaR::central_difference} for a better estimate of velocity. If you wish to re-create this function entirely then use \code{tadaR::first_difference}
}
