% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAudioLong.R
\name{getAudioLong}
\alias{getAudioLong}
\title{Obtains audio from read_tada object and returns in long format}
\usage{
getAudioLong(d)
}
\arguments{
\item{d}{tibble produced by read_tada function}
}
\value{
long format object with columns Sample/Time/Variable/Value
}
\description{
Unnests read_tada audio object into long format object
the output of this can be attached to an unnested TADA object (output of unnest_tada) that has \emph{then} been converted to long format and has the columns Sample/Time/Variable
unnest_tada does not work for obtaining audio due to sampling rate differences between audio and articulatory data
}
