% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getWinRanges.R
\name{getWinRanges}
\alias{getWinRanges}
\alias{getWinRanges,GRangesList-method}
\title{Get the elastic sliding window ranges of the smoothed DAR metric}
\usage{
getWinRanges(dar, ...)

\S4method{getWinRanges}{GRangesList}(dar, extendEdges = FALSE)
}
\arguments{
\item{dar}{A GRangesList with ranges representing single base positions
and metadata columns containing the smoothed DAR values.}

\item{...}{Not used}

\item{extendEdges}{A logical specifying if ranges at the edges of each
chromosome should be extended to cover the entire chromosome.
Useful for assigning DAR values to genomic features that exist
at the 5' or 3' edges of the chromosome}
}
\value{
A GRangesList with ranges that represent the elastic sliding windows
used to smooth the DAR metric.
Each element of the list represents a comparison between two sample groups.
}
\description{
Replace the ranges associated with single base position raw
DAR values to the elastic sliding window ranges associated with the smoothed
DAR values
}
\examples{
fl <- system.file("extdata", "chr1.vcf.bgz", package="darr")
genotypes <- readGenotypes(fl)
groups <- list(
    group1 = c("S2", "S7", "S9", "S10", "S19", "S20"),
    group2 = c("S3", "S6", "S11", "S12", "S15", "S16", "S18")
)
counts <- countAlleles(genotypes, groups)
props <- countsToProps(counts)
contrasts <- matrix(
    data = c(1, -1),
    dimnames = list(
        Levels = c("group1", "group2"),
        Contrasts = c("group1v2")
    )
)
dar <- dar(props, contrasts)
darSmooth <- smoothDar(dar, winSize = 5)
getWinRanges(darSmooth)

}
