% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tada.R
\name{process_tada}
\alias{process_tada}
\title{Process TADA simulation data (smoothing, calculate velocity and tangential velocity)}
\usage{
process_tada(
  d,
  filter_type = "low",
  filter_order = 5,
  filter_cutoffs = 20,
  sampling_rate = 200,
  abs = TRUE
)
}
\arguments{
\item{d}{tibble produced by unnest_tada function}

\item{filter_type}{Filter type; either "low", "high", "pass" or "stop" (default is "low")}

\item{filter_order}{Order of the filter (default is 5 = 5th-order)}

\item{filter_cutoffs}{Cut-off frequency for low-pass filter (default is 20 Hz; for pass or stop filter cutoffs should be of length 2)}

\item{sampling_rate}{Sampling rate of TADA data (default is 200 Hz)}

\item{abs}{return absolute values for _vel columns (default is TRUE; _tvel returns the magnitude so is inherently absolute-valued even if abs=FALSE)}
}
\value{
Returns the input object with the addition of _filt, _vel and _tvel variables for model articulator columns
}
\description{
Smooths all model articulator variables and calculates absolute-valued velocity in x- and y-dimensions, as well as tangential velocity in x/y plane, for model articulatory variables. Smoothing and velocity calculation is done by calling tadaR::butter_filt and tadaR::get_velocity, which can also be used externally to this function.
}
