% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tada.R
\name{read_tada}
\alias{read_tada}
\title{Converts TADA matlab file to a nested tibble}
\usage{
read_tada(matfile)
}
\arguments{
\item{matfile}{quoted filepath to .mat file generated by TADA via the 'saveMView' option}
}
\value{
tibble with a list for each signal type
}
\description{
Reads .mat files produced by Task Dynamics Application
}
