% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/readGenotypes.R
\name{readGenotypes}
\alias{readGenotypes}
\alias{readGenotypes,character-method}
\title{Read genotypes from a VCF file}
\usage{
readGenotypes(file, ...)

\S4method{readGenotypes}{character}(file, ...)
}
\arguments{
\item{file}{The file path of a VCF file containing genotype data}

\item{...}{Passed to \code{\link[VariantAnnotation:readVcf-methods]{VariantAnnotation::readVcf()}}}
}
\value{
tmp
}
\description{
Extract genotypes from a VCF file into a GRanges object
}
\details{
Extract genotype information from the GT field
}
\examples{
library(VariantAnnotation)
fl <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
readGenotypes(fl, genome = "hg19")

}
