% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/smoothDar.R
\name{smoothDar}
\alias{smoothDar}
\alias{smoothDar,GRangesList-method}
\title{Smooth the Differential Allelic Representation metric}
\usage{
smoothDar(dar, ...)

\S4method{smoothDar}{GRangesList}(dar, winSize = 5)
}
\arguments{
\item{dar}{A GRangesList containing DAR values at each overlapping range
between two sample groups.
Each element of the list represents a comparison between two sample groups}

\item{...}{Not used}

\item{winSize}{integer specifying the number of ranges to include in the
elastic sliding window used for smoothing the DAR metric}
}
\value{
A GRangesList containing smoothed DAR values calculated as a rolling
average over a specified number of ranges.
Each element of the list represents a comparison between two sample groups.
}
\description{
Add a metadata column alongside raw DAR values containing
smoothed DAR values that are calculated using an elastic sliding window
approach
}
\examples{
fl <- system.file("extdata", "chr1.vcf.gz", package="darr")
genotypes <- readGenotypes(fl)
groups <- list(
    group1 = c("S2", "S7", "S9", "S10", "S19", "S20"),
    group2 = c("S3", "S6", "S11", "S12", "S15", "S16", "S18")
)
counts <- countAlleles(genotypes, groups)
props <- countsToProps(counts)
contrasts <- matrix(
    data = c(1, -1),
    dimnames = list(
        Levels = c("group1", "group2"),
        Contrasts = c("group1v2")
    )
)
dar <- dar(props, contrasts)
smoothDar(dar, winSize = 5)

}
