% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/unphaseGT.R
\name{unphaseGT}
\alias{unphaseGT}
\alias{unphaseGT,matrix-method}
\alias{unphaseGT,data.frame-method}
\title{Unphase genotypes}
\usage{
unphaseGT(gt)

\S4method{unphaseGT}{matrix}(gt)

\S4method{unphaseGT}{data.frame}(gt)
}
\arguments{
\item{gt}{A matrix or data.frame containing only genotype information}
}
\value{
A \link[base]{matrix} containing unphased genotype calls
}
\description{
Remove phasing information from genotype calls
}
\details{
Phasing information is not required for a simple DAR analysis.
Removing this enables easy counting of alleles from genotype calls.
}
\examples{
library(VariantAnnotation)
fl <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- readVcf(fl, "hg19")
gt <- geno(vcf)$GT
unphaseGT(gt)

}
