% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unphaseGT.R
\name{unphaseGT}
\alias{unphaseGT}
\title{Unphase genotypes}
\usage{
unphaseGT(gt)
}
\arguments{
\item{gt}{matrix or data.frame containing genotype information}
}
\value{
A \link[base]{data.frame} containing unphased genotype calls
}
\description{
Remove phasing information from genotype calls
}
\details{
Phasing information is not required for a simple DAR analysis.
Removing this enables easy counting of alleles from genotype calls.
}
\examples{
library(VariantAnnotation)
fl <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- readVcf(fl, "hg19")
gt <- geno(vcf)$GT
unphaseGT(gt)

}
