% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/convertRanges.R
\name{convertRanges}
\alias{convertRanges}
\alias{convertRanges,GRangesList-method}
\title{Convert DAR origin ranges to DAR region ranges, or vice versa}
\usage{
convertRanges(dar, extendEdges = FALSE)

\S4method{convertRanges}{GRangesList}(dar, extendEdges = FALSE)
}
\arguments{
\item{dar}{A GRangesList with ranges representing single nucleotide (origin)
positions}

\item{extendEdges}{A logical specifying if region DAR ranges at the edges of
each chromosome should be extended to cover the entire chromosome.
This argument is only considered when converting from origin DAR ranges to
region DAR ranges.
Useful for downstream assignment of DAR values to genomic features that
exist at the 5' or 3' edges of the chromosome, which would have otherwise
been missed}
}
\value{
A GRangesList with ranges that represent either DAR regions or
DAR origins, depending on the ranges of the input object.
}
\description{
Convert the ranges element associated with origin DAR values
to ranges associated with the region DAR values.
This function can also be used to revert back to the original object
containing origin ranges if desired
}
\examples{
fl <- system.file("extdata", "chr1.vcf.bgz", package="darr")
genotypes <- readGenotypes(fl)
groups <- list(
    group1 = c("S2", "S7", "S9", "S10", "S19", "S20"),
    group2 = c("S3", "S6", "S11", "S12", "S15", "S16", "S18")
)
counts <- countAlleles(genotypes, groups)
props <- countsToProps(counts)
contrasts <- matrix(
    data = c(1, -1),
    dimnames = list(
        Levels = c("group1", "group2"),
        Contrasts = c("group1v2")
    )
)
dar <- dar(props, contrasts, winSize = 5)
## Convert ranges to regions associated with dar_region values
convertRanges(dar)
## Extend the outer regions the cover the entire chromosome
convertRanges(dar, extendEdges = TRUE)

## Convert back to origin ranges associated with dar_origin values
darRegions <- convertRanges(dar)
convertRanges(darRegions)

}
