% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/countsToProps.R
\name{countsToProps}
\alias{countsToProps}
\alias{countsToProps,GRangesList-method}
\title{Convert allele counts to proportions}
\usage{
countsToProps(counts, ...)

\S4method{countsToProps}{GRangesList}(counts, filter = TRUE)
}
\arguments{
\item{counts}{A GRangesList containing a summary of allele counts at each
range.
Each element of the list represents a distinct sample group}

\item{...}{Not used}

\item{filter}{A \code{logical} specifying if ranges should be filtered using
default criteria.
If filtering is not required, or if performed manually with
\code{\link[=filterLoci]{filterLoci()}}, set this to \code{FALSE}}
}
\value{
A GRangesList containing a summary of normalised allele counts
(i.e. as proportions) at each range.
Each element of the list represents a distinct sample group
}
\description{
Normalise allele-level counts across samples by converting to a
proportion of total alleles in all samples.
Optionally remove ranges using the default filtering criteria described
in \code{\link[=filterLoci]{filterLoci()}}
}
\examples{
fl <- system.file("extdata", "chr1.vcf.bgz", package="darr")
genotypes <- readGenotypes(fl)
groups <- list(
    group1 = c("S2", "S7", "S9", "S10", "S19", "S20"),
    group2 = c("S3", "S6", "S11", "S12", "S15", "S16", "S18")
)
counts <- countAlleles(genotypes, groups)
countsToProps(counts)

}
