% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dar.R
\name{dar}
\alias{dar}
\alias{dar,GRangesList,matrix-method}
\title{Calculate Differential Allelic Representation (DAR)}
\usage{
dar(props, contrasts, ...)

\S4method{dar}{GRangesList,matrix}(props, contrasts)
}
\arguments{
\item{props}{A GRangesList containing a summary of normalised allele counts
(i.e. as proportions) at each range.
Each element of the list represents a distinct sample group}

\item{contrasts}{A contrast matrix specifying which sample groups to
to calculate DAR between.
Each column must represent a single contrast, and rows represent the levels
(i.e. sample groups) to be contrasted.
The two levels involved with each contrast should be specified with
\code{1} and \code{-1}}

\item{...}{Not used}
}
\value{
A GRangesList containing DAR values at each overlapping range
between the contrasted sample groups.
Each element of the list represents a comparison between two sample groups.
}
\description{
Calculate DAR between two sample groups
}
\examples{
fl <- system.file("extdata", "chr1.vcf.gz", package="darr")
genotypes <- readGenotypes(fl)
groups <- list(
    group1 = c("S2", "S7", "S9", "S10", "S19", "S20"),
    group2 = c("S3", "S6", "S11", "S12", "S15", "S16", "S18")
)
counts <- countAlleles(genotypes, groups)
props <- countsToProps(counts)
contrasts <- matrix(
    data = c(1, -1),
    dimnames = list(
        Levels = c("group1", "group2"),
        Contrasts = c("group1v2")
    )
)
dar(props, contrasts)

}
