% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dar.R
\name{dar}
\alias{dar}
\alias{dar,GRangesList,matrix-method}
\title{Calculate Differential Allelic Representation (DAR)}
\usage{
dar(props, contrasts, winSize = 5)

\S4method{dar}{GRangesList,matrix}(props, contrasts, winSize = 5)
}
\arguments{
\item{props}{A GRangesList containing a summary of normalised allele counts
(i.e. as proportions) at each range.
Each element of the list represents a distinct sample group}

\item{contrasts}{A contrast matrix specifying which sample groups to
to calculate DAR between.
Each column must represent a single contrast, and rows represent the levels
(i.e. sample groups) to be contrasted.
The two levels involved with each contrast should be specified with
\code{1} and \code{-1}}

\item{winSize}{integer specifying the number of ranges to include
in the elastic sliding window used for averaging DAR values within a region.
Must be an odd integer in order to incorporate the origin locus and an
equal number of loci either side}
}
\value{
A GRangesList containing DAR values at each overlapping range
between the contrasted sample groups.
Two types of DAR values are reported in the metadata columns of each GRanges
object:
\itemize{
\item dar_origin: The raw DAR values calculated at single nucleotide positions
(the origin) between sample groups.
\item dar_region: The mean of raw DAR values in a region surrounding the origin.
The size of the region is controlled using the \code{winSize} argument, which
establishes an elastic sliding window to average the specified number
of dar_origin values.
}

Each element of the list represents a single contrast defined in the
input contrast matrix
}
\description{
Calculate DAR between two sample groups
}
\details{
DAR is calculated as the Euclidean distance between the allelic
proportions (i.e. proportion of As, Cs, Gs and Ts) of two sample groups at
a single nucleotide locus, scaled such that all values range inclusively
between 0 and 1.
A DAR value of 0 represents identical allelic representation between the
two sample groups, while a DAR value of 1 represents complete diversity
}
\examples{
fl <- system.file("extdata", "chr1.vcf.bgz", package="darr")
genotypes <- readGenotypes(fl)
groups <- list(
    group1 = c("S2", "S7", "S9", "S10", "S19", "S20"),
    group2 = c("S3", "S6", "S11", "S12", "S15", "S16", "S18")
)
counts <- countAlleles(genotypes, groups)
props <- countsToProps(counts)
contrasts <- matrix(
    data = c(1, -1),
    dimnames = list(
        Levels = c("group1", "group2"),
        Contrasts = c("group1v2")
    )
)
dar(props, contrasts, winSize = 5)

}
